Remove HexConverter and fix namespaces
