+++ src/TgMiniAppAuth/AuthContext/TelegramAuthorizationContextValidator.cs
@@ -1,7 +1,8 @@
+using System.Runtime.CompilerServices;
 using System.Security.Cryptography;
 using System.Text;
 
-namespace TgMiniAppAuth.Authorization;
+namespace TgMiniAppAuth.AuthContext;
 
 /// <summary>
 /// Telgram mini app auth context
@@ -14,9 +15,8 @@ internal class TelegramAuthorizationContextValidator : ITelegramAuthorizationCon
     private static readonly byte[] WebAppDataBytes = "WebAppData"u8.ToArray();
 
     private const int StackAllocationThreshold = 1024;
-    
-    /// <inheritdoc />
 
+    /// <inheritdoc />
     /// <summary>
     /// Check that hash value valid sign of all pairs except 'hash=*' of <see cref="WebAppDataBytes"/> with the token of the telegram bot. 
     /// </summary>
@@ -303,10 +303,10 @@ internal class TelegramAuthorizationContextValidator : ITelegramAuthorizationCon
             {
                 if (input[pos + 1] == 'u' && pos < count - 5)
                 {
-                    int h1 = HexConverter.FromChar(input[pos + 2]);
-                    int h2 = HexConverter.FromChar(input[pos + 3]);
-                    int h3 = HexConverter.FromChar(input[pos + 4]);
-                    int h4 = HexConverter.FromChar(input[pos + 5]);
+                    int h1 = FromChar(input[pos + 2]);
+                    int h2 = FromChar(input[pos + 3]);
+                    int h3 = FromChar(input[pos + 4]);
+                    int h4 = FromChar(input[pos + 5]);
 
                     if ((h1 | h2 | h3 | h4) != 0xFF)
                     {
@@ -321,8 +321,8 @@ internal class TelegramAuthorizationContextValidator : ITelegramAuthorizationCon
                 }
                 else
                 {
-                    int h1 = HexConverter.FromChar(input[pos + 1]);
-                    int h2 = HexConverter.FromChar(input[pos + 2]);
+                    int h1 = FromChar(input[pos + 1]);
+                    int h2 = FromChar(input[pos + 2]);
 
                     if ((h1 | h2) != 0xFF)
                     {
@@ -350,6 +350,34 @@ internal class TelegramAuthorizationContextValidator : ITelegramAuthorizationCon
         length = helper.BytesLength;
         helper.WriteString(decoded);
     }
+
+
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    public static int FromChar(int c)
+    {
+        return (c >= CharToHexLookup.Length) ? 0xFF : CharToHexLookup[c];
+    }
+
+    /// <summary>Map from an ASCII char to its hex value, e.g. arr['b'] == 11. 0xFF means it's not a hex digit.</summary>
+    public static ReadOnlySpan<byte> CharToHexLookup =>
+    [
+        0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, // 15
+        0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, // 31
+        0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, // 47
+        0x0, 0x1, 0x2, 0x3, 0x4, 0x5, 0x6, 0x7, 0x8, 0x9, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, // 63
+        0xFF, 0xA, 0xB, 0xC, 0xD, 0xE, 0xF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, // 79
+        0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, // 95
+        0xFF, 0xa, 0xb, 0xc, 0xd, 0xe, 0xf, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, // 111
+        0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, // 127
+        0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, // 143
+        0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, // 159
+        0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, // 175
+        0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, // 191
+        0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, // 207
+        0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, // 223
+        0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, // 239
+        0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF // 255
+    ];
 
     // Microsoft: System.Web.HttpUtility.UrlEncoder
     // Internal class to facilitate URL decoding -- keeps char buffer and byte buffer, allows appending of either chars or bytes
